Consolidate deploy scripts
