--- protocol/scripts/deploys/mainnet/15-deploy-team-payments.ts
@@ -1,94 +1,0 @@
-import '@nomiclabs/hardhat-ethers';
-import { BigNumber } from 'ethers';
-import { ethers, network } from 'hardhat';
-import { AcceleratedExitQueue__factory, AMMWhitelist, AMMWhitelist__factory, ExitQueue, ExitQueue__factory, TempleERC20Token__factory, TempleFraxAMMOps, TempleFraxAMMOps__factory, TempleFraxAMMRouter, TempleFraxAMMRouter__factory, TempleTeamPayments, TempleTeamPayments__factory, TempleUniswapV2Pair, TempleUniswapV2Pair__factory, TreasuryManagementProxy__factory } from '../../../typechain';
-import { deployAndMine, DeployedContracts, DEPLOYED_CONTRACTS, expectAddressWithPrivateKey, fromAtto, mine, toAtto } from '../helpers';
-
-async function main() {
-  expectAddressWithPrivateKey();
-  const [owner] = await ethers.getSigners();
-
-  let DEPLOYED: DeployedContracts;
-
-  if (DEPLOYED_CONTRACTS[network.name] === undefined) {
-    console.log(`No contracts configured for ${network.name}`)
-    return;
-  } else {
-    DEPLOYED = DEPLOYED_CONTRACTS[network.name];
-  }
-
-  const templeTeamPaymentsFactory = new TempleTeamPayments__factory(owner);
-  const templeTeamFixedPayments: TempleTeamPayments = await deployAndMine(
-    'TEMPLE_TEAM_FIXED_PAYMENTS', templeTeamPaymentsFactory, templeTeamPaymentsFactory.deploy,
-    DEPLOYED.TEMPLE,
-    13140000, // 5 months in seconds
-    Math.round(Date.now()/1000 - 5256000) // 2 months ago
-  )
-
-  const allocs : [string, number][] = [
-    ["0x2df7fc713668381585d13C8580A1028F9fe02296",76472.0876879939],
-["0x4CC67b60176C62aB316F0AD06D141b15e06f073e",178434.871271986],
-["0x39c4225C08B89eC411b730d39b2CCd927e56C764",497068.56997196],
-["0xa4ee464cE1A9058B44a8831085D1413eE235D10F",1159826.66326791],
-["0xB792fFF55eC8FC7Fde4e36E32d3f5d227d7eE363",91766.5052255927],
-["0x67604d7a1895Bfb76750661FD468706082c77dFE",214121.845526383],
-["0x4eB8259d2FD604BFba4F0aDfFacEAa12D3e4c987",305888.350751976],
-["0x9a675CaEDE7Bf278424102ccC1861576289732C0",713739.485087943],
-["0x522689D6E11d09773133Cb62c80ed40aa2722D7b",267652.306907979],
-["0x07c9cdf5BAD644A6CB5A1eb36627f6603E18E095",624522.04945195],
-["0x135EA0161c5047479e7EDCF0A9c5fdd357ade9B8",50981.3917919959],
-["0x2d9D668999090f56ED4Ca46738E51E56885f67B4",25490.695895998],
-["0xf9dd7eA0fc45148e89c817e7956Ab141652Ca9a8",30588.8350751976],
-["0x31ff22aa6f49DFC971Eb9266Ee98FEBE161d5f00",7647.20876879939],
-["0x735da150245ca19801b65fE4F4616A788d6712F2",40785.1134335967],
-["0x6127cb39Ac8e6066C469aCE0edcC3506feaAbF94",12745.347947999],
-["0xcB8E497B186932DBaDF55e65A4225173F0d35222",5098.13917919959],
-["0xD7E2622dFe0D8e4fC19fd3DF34BDCab6B6BBa383",15294.4175375988],
-["0xD07449b3D4aD58CfB55B68FbAE783C71d847C4a3",10196.2783583992],
-["0x3FB8C0F39D9b33cE2444e369f3CED0090c9a309A",10196.2783583992],
-["0x7158A57df50B72093127856F0448D6f0B4D14072",5098.13917919959],
-["0xEEf8D69C6468DE515805A05929f6F11C678C0cF4",40785.1134335967],
-["0x35093a15f7310459821E597b074443b6aFC3D8C2",50981.3917919959],
-["0x70B953B65945fc5Fe4B0142B73EEd1cf1D54277D",40785.1134335967],
-["0xF2E1A88af3D6cDdD60B9c43CeE53d655cBa40b2c",12745.347947999],
-["0xE0bb475Ef181168dFF18D2BD7d1C147d4dc502F9",25490.695895998],
-["0xc234c54F6927aC30c6a0Fc22f3e2A54c08fd4C50",9589.35702754209],
-["0x73E29Bb563592610a4DdDfbeEFC0259E3A27e410",13595.0378111989],
-["0x02DdD68E1A4bba3C263718A381C03BAD4a97beC1",6190.59757474236],
-["0x82eAa1b159895cba666615CAe2C24A813465722d",19664.2511197699],
-["0x1AF888B451971EB74Cd2FC203F49AE71543Ff75A",14444.7276743988],
-["0x8caF5ad1c534c8b5B9801401cbF5bD10872cafE9",18086.2556595414],
-["0x33cB0C602d9D2965c5538731bAB28F122988f74E",14869.5726059988],
-["0x5819a28960709e08c4E5c82Cd7027Eb11A079409",9953.50982605635],
-["0x5edef4983f2f5dC70742325b0FB7FE03287cf7E5",7889.97730114223],
-["0x640307c621C7B2a9252407Db87c30C31aC28a66A",13170.1928795989],
-["0xd53Db76369fBb6584E0BeC45d4e8D62eB46E5700",10924.5839554277],
-["0xBF0AB0436150c1928ee16686A0dB560E9d5c581F",14869.5726059988],
-["0x5dfD6240980D53F9B0A301d2A315B9E670Ef6516",13170.1928795989],
-["0x36cf80D36fDA4530777aeD39ae7f398E38fbb25f",9346.58849519925],
-["0xc807bfA8007D68C45D16c01db488513c0629DB50",10439.046890742],
-["0x9545D4a659EA0815Ee65a12A41bE73728D9942E1",9710.74129371351],
-["0xC66b63BEDDB456dCa88Cc72AAa4019f0449D0Fd0",6797.51890559946],
-["0x49f4F69e107cF1010869CB2Cd99e945ACd5BD89E",19785.6353859413],
-["0x5EFC85DfD18ce35427418eEe3980E8eA15F4aa78",5947.82904239953],
-["0xF3509061CE0F0D2f2f11D3825be2d5F18d0f88E0",14080.5748758846],
-["0x717a41b065Ad00c9A91DB61E865Ecdd8Ef22Bc49",17357.9500625129],
-["0x08762b6719D840Ff0f1B308F9406002e105bAb6e",5098.13917919959],
-  ]
-
-  await mine(templeTeamFixedPayments.setAllocations(
-    allocs.map(a => a[0]),
-    allocs.map(a => toAtto(a[1])),
-  ));
-
-  // await templeTeamFixedPayments.transferOwnership(DEPLOYED.MULTISIG)
-}
-
-// We recommend this pattern to be able to use async/await everywhere
-// and properly handle errors.
-main()
-  .then(() => process.exit(0))
-  .catch(error => {
-    console.error(error);
-    process.exit(1);
-  });
