+++ protocol/scripts/deploys/mainnet/team-payments/json/epoch2.json
@@ -1,0 +1,76 @@
+[
+  ["0x522689d6e11d09773133cb62c80ed40aa2722d7b", 125000],
+  ["0x07c9cdf5BAD644A6CB5A1eb36627f6603E18E095", 125000],
+  ["0x4D1a31404c6B14dA50724945f6cd11d0864f272a", 86111.1111111111],
+  ["0xdF088b3A908748d9b9b57a186E769Ff6c0FF1BFC", 86111.1111111111],
+  ["0x8706Db8AcbE6295c3E769a717B815B994CcC6D14", 86111.1111111111],
+  ["0x9Dd22E2CaaEfDadE67D82C26a0f3f5676f1884cf", 86111.1111111111],
+  ["0xFD6E8Bf70beC613eAdb63a7E543FA13f0EDb375C", 86111.1111111111],
+  ["0x06C12cD35604ED9006E7bD11D6FFe1E9126BABf0", 86111.1111111111],
+  ["0x4ce17ae9fb4aa0178ec2f8eea3f6167ae511f14c", 55555.5555555556],
+  ["0x9a675CaEDE7Bf278424102ccC1861576289732C0", 55555.5555555556],
+  ["0x213221b48359186D72Bf47D9f03eAb10B687c698", 55555.5555555556],
+  ["0xb792fff55ec8fc7fde4e36e32d3f5d227d7ee363", 75000],
+  ["0x67604d7a1895Bfb76750661FD468706082c77dFE", 75000],
+  ["0xf9dd7ea0fc45148e89c817e7956ab141652ca9a8", 100000],
+  ["0x547ABB8aDC1C6505c467e9D27a647f0037883E71", 75000],
+  ["0x70b953b65945fc5fe4b0142b73eed1cf1d54277d", 25833.3333333333],
+  ["0xe0bb475ef181168dff18d2bd7d1c147d4dc502f9", 45833.3333333333],
+  ["0x36afcc3f7403a7059ba034eee01688859880cd0b", 45833.3333333333],
+  ["0xf2e1a88af3d6cddd60b9c43cee53d655cba40b2c", 45833.3333333333],
+  ["0x35093a15f7310459821e597b074443b6afc3d8c2", 33333.3333333333],
+  ["0x491c7d47c8bc4535acff5b38e386bcd89ab6ce28", 45833.3333333333],
+  ["0x760a8e15155a9907cb1a87ef9c160ef7360528fe", 45833.3333333333],
+  ["0xeef8d69c6468de515805a05929f6f11c678c0cf4", 33333.3333333333],
+  ["0x135ea0161c5047479e7edcf0a9c5fdd357ade9b8", 16666.6666666667],
+  ["0x54120e6f4379a525acc5ae7dbbf6758c271b8b15", 33333.3333333333],
+  ["0x735da150245ca19801b65fe4f4616a788d6712f2", 37500],
+  ["0xe345e9605391c13fd2502e7141d54b752aadb184", 33333.3333333333],
+  ["0x0f511fe5dcdb54b0b19a5a2a57e76977ae728c15", 25000],
+  ["0xaa64cfdc7188246e12e023260cf05a21e13567a4", 20833.3333333333],
+  ["0x3fb8c0f39d9b33ce2444e369f3ced0090c9a309a", 20833.3333333333],
+  ["0xd959d4c2d90fd2f291a71bf923da24e374e1ef07", 16666.6666666667],
+  ["0xd7e2622dfe0d8e4fc19fd3df34bdcab6b6bba383", 25000],
+  ["0x8a0b19dd5072c7584d6cce8e1e679901c6d3f6c5", 33333.3333333333],
+  ["0x3c1e397d428281f156f8fa733ba160faa2bab946", 33333.3333333333],
+  ["0x1c079e7fc9d8a75e4055cbf6a83f9bbfa13f7103", 20833.3333333333],
+  ["0x8caf5ad1c534c8b5b9801401cbf5bd10872cafe9", 16666.6666666667],
+  ["0x90e23d451a716bf25987f055e3e57c1384337fde", 20833.3333333333],
+  ["0x93c0824cdeb2675317f05d4db08c1f90fdb1d8ea", 25000],
+  ["0x5dfd6240980d53f9b0a301d2a315b9e670ef6516", 20833.3333333333],
+  ["0xcd57b2cdc5acabb8cd3ab08a40e4eedaaf3ee762", 16666.6666666667],
+  ["0x8eb4d0651924743b80743628f59869c4ecf934e2", 16666.6666666667],
+  ["0xcc66308c7ff2c5a317d384c770f6a5b7b5d97b0d", 12500],
+  ["0x4e67a556cc3d397996020077393ad4b2e0b71020", 16666.6666666667],
+  ["0xd07449b3d4ad58cfb55b68fbae783c71d847c4a3", 8333.33333333333],
+  ["0xcb8e497b186932dbadf55e65a4225173f0d35222", 12500],
+  ["0x5819a28960709e08c4e5c82cd7027eb11a079409", 6250],
+  ["0xaeabd740c58921fe0c2d0dc8f7f202f345d7d7fb", 12500],
+  ["0xb122f96f40203ed018e0f17126dddd0d79855c8a", 12500],
+  ["0x519c187ac6ee6cab5c6fad22344895c84487b85e", 6250],
+  ["0x4c3f9ab82cee8c471d856d03b8529dcf7d57237a", 6250],
+  ["0x73e29bb563592610a4dddfbeefc0259e3a27e410", 12500],
+  ["0xb58617F6aEE3a3A3b76A0fcF10C186cbef54CC3c", 12500],
+  ["0x02DdD68E1A4bba3C263718A381C03BAD4a97beC1", 20833.3333333333],
+  ["0x82eaa1b159895cba666615cae2c24a813465722d", 8333.33333333333],
+  ["0x3ac87ed1bedb341f97656d5f24eddef7fc810811", 18750],
+  ["0xae093062a44bf35e55891f996a0f5518c778dad3", 6250],
+  ["0xc20aceda0a6fa2b79a5de9a05186809264fea05d", 18750],
+  ["0x1a856a0e6a63e289582faf9b96e05ab1406e6b07", 6250],
+  ["0xc086c55fdd80418e62e1789d9174aa2a639a6b1f", 20833.3333333333],
+  ["0xd53db76369fbb6584e0bec45d4e8d62eb46e5700", 12500],
+  ["0x45a7c62f2ba4a31d07b28f81bf63c67968b0efb9", 6250],
+  ["0x2df7fc713668381585d13c8580a1028f9fe02296", 25000],
+  ["0x3212db021cd409e2129f1c85b55ab039212d1f52", 12500],
+  ["0xf9b4de4bb735ace677813ac8e24b6530b2d595e9", 12500],
+  ["0x2d9d668999090f56ed4ca46738e51e56885f67b4", 8333.33333333333],
+  ["0xc807bfa8007d68c45d16c01db488513c0629db50", 8333.33333333333],
+  ["0x38eEba350EE79Dc178e62159079c7faD11F15d99", 8333.33333333333],
+  ["0xce05a1d344a7c8172d7b484e7d8b78884a6958ab", 16666.6666666667],
+  ["0x97dcd6d467af9d63fecabe1652f30fd483175ebd", 12500],
+  ["0x9a130433d06434d79005b6befebdc0f65ef2e811", 6250],
+  ["0x08e19c5b9667fa2d49a6709482aa5ac0c4a7b198", 12500],
+  ["0xE77D5F652f4a7aDad42f3ff1f33608C7744c9173", 8333.33333333333],
+  ["0x5993babe0fc029e1de368ed3c831f0dbb914b010", 25000],
+  ["0x9545D4a659EA0815Ee65a12A41bE73728D9942E1", 8333.33333333333]
+]
