+++ protocol/scripts/deploys/mainnet/team-payments/json/epoch8.json
@@ -1,0 +1,36 @@
+{
+  "0x63E7916ee248b9D966C1d53142B09F7630b689Be": 2062,
+  "0x7F637eB7463eF4730898656fD8b39bB10B240569": 12683,
+  "0xF9B4De4bB735ACE677813ac8E24B6530B2D595E9": 1031,
+  "0xF94A14fdEbf75007329E50f0aC9f459fF566B157": 464,
+  "0x228900C0263B6Ac19E66fF491E30c83d3473a2d8": 849,
+  "0xe345e9605391c13Fd2502e7141D54b752AaDB184": 2889,
+  "0x87997a32A5Ba852D95CFC7c803B3b3371B223aE2": 512,
+  "0x62370003cEbacd6415d32375095FFc8AB7238eBC": 644,
+  "0x4e51F68940FaC327B2f018B6b79109031031fe8b": 1031,
+  "0xd82C6E486FBd0Be55fD5589532aF8D69aB29C1AA": 515,
+  "0x8BBF709345fA5f6FDB451f95D398EBbCE1750A51": 696,
+  "0x08e19c5B9667FA2d49A6709482aa5aC0C4a7b198": 644,
+  "0xbCB921901a212E9876DB73fa32C8d92aCB6AAbF8": 1443,
+  "0xe5D74520A77EDaa6Fff050975775213FA01948c9": 2577,
+  "0xfED4FEd3C90C4e93D203D25EF2B148C863b68356": 1031,
+  "0x588282ADA5B9D872d0eBF506c5C1b541e2a60BBa": 4536,
+  "0x10AD3Fc9eE37acCdA9E58439C53B8A564C31eC5b": 3093,
+  "0x832d7aed44ff0396770Be2F2f3603153A8c787E9": 1203,
+  "0x9A6Bd26aA38b43D85d98EA03Fb734D24585b743D": 2969,
+  "0x2c90982BEc7A5aD00A65ce87C4F62263A3055cBB": 3436,
+  "0x06e6F9DE62afa6F2Ca092Fe17fF57dAc3Ae30db1": 1443,
+  "0x7f068d2890bD5ce602DCB9d040515306d067cA02": 515,
+  "0x73E29Bb563592610a4DdDfbeEFC0259E3A27e410": 103,
+  "0x387a652c87A2571b93EF4dECc61BEB74A5970582": 515,
+  "0x294A29Ad8842E7234EE16B1b16f1F98747B715Eb": 52,
+  "0x3f0e5DbC6093E8773F14f0186FFE32F733C34ef7": 206,
+  "0xDC884904D07d201c54D1aa5D0AD509E4eE289eAe": 206,
+  "0x8CC8bA612168B5A8cBefb977D7DE7375A61A1009": 8660,
+  "0x779731032c24356199a7baaFae20f4271c0Ccc80": 1134,
+  "0x9743b92460fC79F070D2059D6C533bE84A1dd493": 8247,
+  "0x0AAA67dcbBD86dBd1F0c260F50fb9FB3A941138A": 1598,
+  "0x8D470e6c39d3106c9D4239b6338451c649B8355A": 2474,
+  "0x056c7b90255cA73EF674F75d9653bB5d947c4C9E": 82,
+  "0x71e41D0dFeA7ca196c7B104F01EfFd1102af9694": 23408
+}
